working progress: advances on separete rrDisplay functionalities
