+++ src/app/Score.ts
@@ -1,10 +1,12 @@
 import Root from './Root';
 import Claim from './Claim';
+import Operation from './Operation';
 
 
 ﻿export default class Score {
     claimId: string;
-
+    selectedScore: Score;
+    operation: Operation = new Operation();
     /**  */
     confidencePro: number;
     confidenceCon: number;
@@ -53,6 +55,14 @@ import Claim from './Claim';
 
     isMain: boolean;
     isEditing: boolean;
+
+    select(score: Score, e: Event): void {
+        if (score != this.selectedScore) {
+            this.selectedScore = score;
+            this.operation.setDisplayState();
+            this.operation.update();
+        }
+    }
 }
 
 type DisplayState = "newClaim" | "notSelected" | "parent" | "ancestor" | "selected"| "selected editing" | "child" | "sibling";
