+++ src/app/rrDisplay.ts
@@ -1,3 +1,55 @@
+//Should be this class
+
+// this.calculate();
+  // this.settleIt.calculate
+
+// this.setDisplayState();
+  // this.clearDisplayState();
+  // this.setDisplayStateLoop();
+
+// this.update();
+  // this.renderNode();
+    // this.update();
+    // this.renderNode();
+
+// this.renderNode();
+  // this.selectScore();
+  // this.update();
+  // this.renderNode();
+
+// this.clearDisplayState();
+
+// this.initRr();
+  // this.setDisplayState();
+    // this.clearDisplayState();
+    // this.setDisplayStateLoop();
+  // this.calculate();
+    // this.settleIt.calculate
+
+// this.selectScore();
+  // this.setDisplayState();
+  // this.update();
+    // this.renderNode();
+      // this.renderNode();
+      // this.update();
+
+// this.updateClaim();
+  // this.calculate();
+  // this.update();
+
+// this.removeClaim();
+  // this.calculate();
+  // this.setDisplayState();
+  // this.update();
+
+// this.addClaim();
+  // this.calculate();
+  // this.setDisplayState();
+  // this.update();
+
+// this.editClaim();
+  // this.update();
+
 declare const require: any;
 
 type WhichCopy = "original" | "local" | "suggestion";
@@ -137,45 +189,45 @@ export default class RRDisplay {
         }
     }
 
-    clearDisplayState(): void {
-        for (let scoreId in this.scores) {
-            if (this.scores.hasOwnProperty(scoreId)) {
-                this.scores[scoreId].displayState = "notSelected";
-            }
-        }
-    }
-
-    setDisplayState(): void {
-        this.clearDisplayState();
-        this.setDisplayStateLoop(this.mainScore);
-    }
-
-    setDisplayStateLoop(score: Score): void {
-        if (score == this.selectedScore)
-            score.displayState = "selected";
-
-        for (let childId of this.claims[score.claimId].childIds) {
-            let childScore = this.scores[childId];
-            //process the children first/
-            this.setDisplayStateLoop(childScore);
-
-            if (childScore == this.selectedScore) {
-                score.displayState = "parent";
-                //Set Siblings
-                for (let siblingId of this.claims[score.claimId].childIds) {
-                    let siblingScore = this.scores[siblingId];
-                    if (siblingScore.displayState != "selected")
-                        siblingScore.displayState = "sibling";
-                }
-            }
-
-            if (childScore.displayState == "ancestor" || childScore.displayState == "parent")
-                score.displayState = "ancestor";
-
-            if (score == this.selectedScore)
-                childScore.displayState = "child";
-        }
-    }
+    // clearDisplayState(): void {
+    //     for (let scoreId in this.scores) {
+    //         if (this.scores.hasOwnProperty(scoreId)) {
+    //             this.scores[scoreId].displayState = "notSelected";
+    //         }
+    //     }
+    // }
+    //
+    // setDisplayState(): void {
+    //     this.clearDisplayState();
+    //     this.setDisplayStateLoop(this.mainScore);
+    // }
+    //
+    // setDisplayStateLoop(score: Score): void {
+    //     if (score == this.selectedScore)
+    //         score.displayState = "selected";
+    //
+    //     for (let childId of this.claims[score.claimId].childIds) {
+    //         let childScore = this.scores[childId];
+    //         //process the children first/
+    //         this.setDisplayStateLoop(childScore);
+    //
+    //         if (childScore == this.selectedScore) {
+    //             score.displayState = "parent";
+    //             //Set Siblings
+    //             for (let siblingId of this.claims[score.claimId].childIds) {
+    //                 let siblingScore = this.scores[siblingId];
+    //                 if (siblingScore.displayState != "selected")
+    //                     siblingScore.displayState = "sibling";
+    //             }
+    //         }
+    //
+    //         if (childScore.displayState == "ancestor" || childScore.displayState == "parent")
+    //             score.displayState = "ancestor";
+    //
+    //         if (score == this.selectedScore)
+    //             childScore.displayState = "child";
+    //     }
+    // }
 
     update(): void {
         // if (!this.settings.noAutoSave)
@@ -224,94 +276,94 @@ export default class RRDisplay {
         return target.split(search).join(replacement);
     };
 
-    renderNode(score: Score, parent?: Score): void {
-        var claim: Claim = this.claims[score.claimId];
-        var wire = hyperHTML.wire(score);
-        if (this.animation.animateNumbers(this.scores)) setTimeout(() => this.update(), 100);
-
-        var result = wire`
-                <li id="${claim.claimId}" class="${
-            score.displayState +
-            (score.isMain ? ' mainClaim' : '') +
-            (this.settings.isEditing && this.selectedScore == score ? ' editing' : '')}">
-                    <div class="claimPad" onclick="${this.selectScore.bind(this, score)}">
-                        <div class="${"claim " + (claim.isProMain ? 'pro' : 'con') + (claim.disabled ? ' disabled ' : '') + (claim.childIds.length > 0 && !score.open ? ' shadow' : '')}" >
-                            <div class="innerClaim">
-                                <span class="${score.generation == 0 ? 'score' : 'points'}" >${
-            (score.generation == 0 ?
-                Math.round(score.animatedWeightedPercentage * 100) + '%' :
-                (score.weightDif != undefined ? Math.floor(Math.abs(score.weightDif)) : ''))
-            }</span>
-
-            <span class="proPoints" >${Math.round(score.weightPro)}</span>
-            <span class="conPoints" >${Math.round(score.weightCon)}</span>
-
-                                ${claim.content}
-                                ${claim.maxConf && claim.maxConf < 100 ? " (maximum confidence set to " + claim.maxConf + "%) " : ""}
-                                <a target="_blank" href="${claim.citationUrl}" onclick="${this.noBubbleClick}">
-                                    <span class="citation">${claim.citation}</span>
-                                </a>
-
-                             </div>
-                        </div>
-
-                        <div class="${"childIndicatorSpace" + (claim.childIds.length == 0 ? '' : ' hasChildren')}">
-                            <div class="${"childIndicator " + (claim.isProMain ? 'pro' : 'con')}">
-                            <div class="childIndicatorInner">
-                            ${score.numDesc} more
-                            </div>
-                            </div>
-                        </div>
-
-                        <div class="claimEditHider">
-                            <div class="claimEditSection">
-                                <input bind="content"  oninput="${this.updateClaim.bind(this, claim)}" ><br>
-                                <input bind="citation" oninput="${this.updateClaim.bind(this, claim)}" ><br>
-                                <input bind="citationUrl" oninput="${this.updateClaim.bind(this, claim)}" ><br>
-                                <label for="maxConf" >Maximum Confidence </label><br/>
-                                <input bind="maxConf" name="maxConf" type="number" oninput="${this.updateClaim.bind(this, claim)}" ><br>
-                                <input type="checkbox" bind="isProMain" onclick="${this.updateClaim.bind(this, claim)}">
-                                <label for="isProMain">Does this claim supports the main claim?</label><br/>
-                                <input type="checkbox" bind="disabled" onclick="${this.updateClaim.bind(this, claim)}">
-                                <label for="disabled">Disabled?</label><br/>
-                                <button onclick="${this.removeClaim.bind(this, claim, parent)}" name="button">
-                                    Remove this claim from it's parent
-                                </button><br/>
-                                ID:${claim.claimId}
-                            </div>
-                        </div>
-
-                        <div class="claimMenuHider">
-                            <div class="claimMenuSection">
-                                <div class="addClaim pro" onclick="${this.addClaim.bind(this, score, true)}">add</div>
-                                <div class="addClaim con" onclick="${this.addClaim.bind(this, score, false)}">add</div>
-                                <div class="editClaimButton" onclick="${this.editClaim.bind(this, score)}">edit</div>
-                            </div>
-                        </div>
-
-                    </div>
-
-                    <ul>${
-            claim.childIds.map((childId, i) => this.renderNode(this.scores[childId], score))
-            }</ul>
-                        </li>`
-
-        if (!wire.default) {
-            wire.default = claim.content;
-            let inputs = result.querySelector('.claimPad').querySelectorAll('input');
-            for (let input of inputs) {
-                var bindName = input.getAttribute("bind")
-                if (bindName) {
-                    if (input.type == "checkbox")
-                        input.checked = claim[bindName];
-                    else
-                        input.value = claim[bindName];
-                }
-            }
-        }
-
-        return result;
-    }
+    // renderNode(score: Score, parent?: Score): void {
+    //     var claim: Claim = this.claims[score.claimId];
+    //     var wire = hyperHTML.wire(score);
+    //     if (this.animation.animateNumbers(this.scores)) setTimeout(() => this.update(), 100);
+    //
+    //     var result = wire`
+    //             <li id="${claim.claimId}" class="${
+    //         score.displayState +
+    //         (score.isMain ? ' mainClaim' : '') +
+    //         (this.settings.isEditing && this.selectedScore == score ? ' editing' : '')}">
+    //                 <div class="claimPad" onclick="${this.selectScore.bind(this, score)}">
+    //                     <div class="${"claim " + (claim.isProMain ? 'pro' : 'con') + (claim.disabled ? ' disabled ' : '') + (claim.childIds.length > 0 && !score.open ? ' shadow' : '')}" >
+    //                         <div class="innerClaim">
+    //                             <span class="${score.generation == 0 ? 'score' : 'points'}" >${
+    //         (score.generation == 0 ?
+    //             Math.round(score.animatedWeightedPercentage * 100) + '%' :
+    //             (score.weightDif != undefined ? Math.floor(Math.abs(score.weightDif)) : ''))
+    //         }</span>
+    //
+    //         <span class="proPoints" >${Math.round(score.weightPro)}</span>
+    //         <span class="conPoints" >${Math.round(score.weightCon)}</span>
+    //
+    //                             ${claim.content}
+    //                             ${claim.maxConf && claim.maxConf < 100 ? " (maximum confidence set to " + claim.maxConf + "%) " : ""}
+    //                             <a target="_blank" href="${claim.citationUrl}" onclick="${this.noBubbleClick}">
+    //                                 <span class="citation">${claim.citation}</span>
+    //                             </a>
+    //
+    //                          </div>
+    //                     </div>
+    //
+    //                     <div class="${"childIndicatorSpace" + (claim.childIds.length == 0 ? '' : ' hasChildren')}">
+    //                         <div class="${"childIndicator " + (claim.isProMain ? 'pro' : 'con')}">
+    //                         <div class="childIndicatorInner">
+    //                         ${score.numDesc} more
+    //                         </div>
+    //                         </div>
+    //                     </div>
+    //
+    //                     <div class="claimEditHider">
+    //                         <div class="claimEditSection">
+    //                             <input bind="content"  oninput="${this.updateClaim.bind(this, claim)}" ><br>
+    //                             <input bind="citation" oninput="${this.updateClaim.bind(this, claim)}" ><br>
+    //                             <input bind="citationUrl" oninput="${this.updateClaim.bind(this, claim)}" ><br>
+    //                             <label for="maxConf" >Maximum Confidence </label><br/>
+    //                             <input bind="maxConf" name="maxConf" type="number" oninput="${this.updateClaim.bind(this, claim)}" ><br>
+    //                             <input type="checkbox" bind="isProMain" onclick="${this.updateClaim.bind(this, claim)}">
+    //                             <label for="isProMain">Does this claim supports the main claim?</label><br/>
+    //                             <input type="checkbox" bind="disabled" onclick="${this.updateClaim.bind(this, claim)}">
+    //                             <label for="disabled">Disabled?</label><br/>
+    //                             <button onclick="${this.removeClaim.bind(this, claim, parent)}" name="button">
+    //                                 Remove this claim from it's parent
+    //                             </button><br/>
+    //                             ID:${claim.claimId}
+    //                         </div>
+    //                     </div>
+    //
+    //                     <div class="claimMenuHider">
+    //                         <div class="claimMenuSection">
+    //                             <div class="addClaim pro" onclick="${this.addClaim.bind(this, score, true)}">add</div>
+    //                             <div class="addClaim con" onclick="${this.addClaim.bind(this, score, false)}">add</div>
+    //                             <div class="editClaimButton" onclick="${this.editClaim.bind(this, score)}">edit</div>
+    //                         </div>
+    //                     </div>
+    //
+    //                 </div>
+    //
+    //                 <ul>${
+    //         claim.childIds.map((childId, i) => this.renderNode(this.scores[childId], score))
+    //         }</ul>
+    //                     </li>`
+    //
+    //     if (!wire.default) {
+    //         wire.default = claim.content;
+    //         let inputs = result.querySelector('.claimPad').querySelectorAll('input');
+    //         for (let input of inputs) {
+    //             var bindName = input.getAttribute("bind")
+    //             if (bindName) {
+    //                 if (input.type == "checkbox")
+    //                     input.checked = claim[bindName];
+    //                 else
+    //                     input.value = claim[bindName];
+    //             }
+    //         }
+    //     }
+    //
+    //     return result;
+    // }
 
     selectScore(score: Score, e: Event): void {
         if (score != this.selectedScore) {
@@ -359,6 +411,7 @@ export default class RRDisplay {
         var index = this.claims[parentScore.claimId].childIds.indexOf(claim.claimId);
         if (index > -1) this.claims[parentScore.claimId].childIds.splice(index, 1);
         this.selectedScore = parentScore;
+
         this.calculate();
         this.setDisplayState();
         this.update();
